check params exist in request and save the status on the store (#41)

+++ server/api/exercise.js
@@ -3,23 +3,27 @@ const Exercise = require('../model/exercise');
 let exercise = {};
 
 const listExercise = (req, res) => {
-    Exercise.find({ '_id': req.query.id })
-    .populate('_id')
-    .exec(function (err, list) {
-        if(err) return res.status(400).send(err);
+    if (req.query.id) {
+        Exercise.find({ '_id': req.query.id })
+        .populate('_id')
+        .exec(function (err, list) {
+            if(err) return res.status(400).send(err);
 
-        res.status(200).json(list);
-    });
+            res.status(200).json(list);
+        });
+    }
 };
 
 const getExercise = (req, res) => {
-    Exercise.find({ 'workout_id': req.query.id })
-    .populate('workout_id')
-    .exec(function (err, exer) {
-        if(err) return res.status(400).send(err);
+    if (req.query.id) {
+        Exercise.find({ 'workout_id': req.query.id })
+        .populate('workout_id')
+        .exec(function (err, exer) {
+            if(err) return res.status(400).send(err);
 
-        res.status(200).json(exer);
-    });
+            res.status(200).json(exer);
+        });
+    }
 };
 
 const createExercise = (req, res) => {
@@ -42,37 +46,41 @@ const createExercise = (req, res) => {
 };
 
 const updateExercise = (req, res) => {
-    Exercise.findById(req.params.exercise_id, (err, exercise) => {
-        if(err) return res.send(err);
+    if (req.params.exercise_id) {
+        Exercise.findById(req.params.exercise_id, (err, exercise) => {
+            if(err) return res.send(err);
 
-        const data = {};
+            const data = {};
 
-        if (req.body.history) {
-            data.history = req.body.history;
-        }
+            if (req.body.history) {
+                data.history = req.body.history;
+            }
 
-        if (req.body.status) {
-            data.status = req.body.status;
-        }
+            if (req.body.status) {
+                data.status = req.body.status;
+            }
 
-        exercise.set(data);
+            exercise.set(data);
 
-        exercise.save((err, exercise) => {
-            if(err) return res.status(400).send(err);
+            exercise.save((err, exercise) => {
+                if(err) return res.status(400).send(err);
 
-            res.json(exercise);
+                res.json(exercise);
+            });
         });
-    });
+    }
 };
 
 const deleteExercise = (req, res) => {
-    Exercise.deleteOne({
-        _id: req.params.exercise_id
-    }, (err, exercise) => {
-        if(err) return res.status(400).send(err);
+    if (req.params.exercise_id) {
+        Exercise.deleteOne({
+            _id: req.params.exercise_id
+        }, (err, exercise) => {
+            if(err) return res.status(400).send(err);
 
-        res.json({ message: 'Exercise Deleted.' });
-    });
+            res.json({ message: 'Exercise Deleted.' });
+        });
+    }
 };
 
 module.exports = {
